+++ PowerShellForFOD/Private/Parse-FODAttribute.ps1
@@ -11,6 +11,13 @@ function Parse-FODAttribute
             id = $Attribute.id
             name = $Attribute.name
             value = $Attribute.value
+            attributeTypeId = $Attribute.attributeTypeId
+            attributeType = $Attribute.attributeType
+            attributeDataTypeId = $Attribute.attributeDataTypeId
+            attributeDataType = $Attribute.attributeDataType
+            isRequired = $Attribute.isRequired
+            isRestricted = $Attribute.isRestricted
+            picklistValues = $Attribute.picklistValues
             Raw = $Attribute
         }
     }