+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# Change Log
+# Power Shell For Fortify On Demand (FOD) Module
+
+## Change Log
 
 All notable changes to this project will be documented in this file.
 
@@ -6,7 +8,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
-## [1.0.0.0] - 2020-04-22
+## [1.0.0.2] - 2020-04-23
+
+### Added
+
+- Initial version with token generation and basic FOS API call infrastructure
+- Query, Add, Update and Delete Releases
+- Query, Add, Update and Delete Users
+- Query Attributes
+
+### Changed
+
+- Add-FODApplication (updated documentation)
+- New-FODApplicationObject (changed HasMicroservices to command line switch)
+- Update-FODApplication (updated documentation)
+- Added USAGE.md documentation
+- Updated README.md
+
+### Removed
+
+## [1.0.0.1] - 2020-04-22
 
 ### Added